Here's my first foray into creating a class Ticker
I've attached the WSBcomments I'm using. This ends up creating a final_list which contains all Ticker objects with count >1. The Ticker object's attributes are the ticker, a list of comments where the ticker is mentioned, the count of how many comments mention the ticker, and a list of the compound sentiment for each comment. To add: positive sentiment and negative sentiment items from the sentiment analyzer.